LayoutTests/crypto/workers/subtle/pbkdf2-import-key-derive-bits-expected.txt

[Worker] Test PBKDF2 deriveBits operation with an imported raw key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/pbkdf2-import-key-derive-bits.js
PASS [Worker] bytesToHexString(derivedKey) is expectedDerivedKey
PASS successfullyParsed is true

TEST COMPLETE
